TypeResolver class added
Completely removed the Reflection.ts class as from the API's standpoint
it was not being utilized. The reason being is that Reflections are
passed on the compilation units, not as some overarching data structure.

Introduced a static class called TypeResolver which is handling Type
qualification at the end of a compilation. The TypeResolver will
resolver types for a single or multiple document type space.